money_transfer/requirements.txt
asgiref==3.6.0
attrs==22.2.0
Babel==2.12.1
Django==4.1.7
django-configurations==2.4
django-money==3.0.0
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
py-moneyed==2.0
pytest==7.2.2
pytest-django==4.5.2
pytz==2023.2
sqlparse==0.4.3
typing_extensions==4.5.0
